Refactor to add some separate derivation logic
